requirements.txt
Django>=5.0.2
djangorestframework>=3.14.0
python-dotenv>=1.0.0
transformers>=4.37.2
torch>=2.2.0
torchaudio>=2.2.0
pyannote.audio==3.1.1
openai-whisper>=20231117
numpy>=1.24.0
soundfile>=0.12.1
librosa>=0.10.1
python-magic>=0.4.27
django-cors-headers>=4.3.1
gunicorn>=21.2.0
whitenoise>=6.6.0
huggingface-hub>=0.20.3
pydub>=0.25.1
psutil>=5.9.0 